+++ src/main/java/tornadofx/App.kt
@@ -4,7 +4,6 @@ import javafx.application.Application
 import javafx.application.Platform
 import javafx.scene.Scene
 import javafx.scene.image.Image
-import javafx.scene.layout.Pane
 import javafx.stage.Stage
 import java.awt.*
 import java.awt.event.MouseAdapter
@@ -21,7 +20,7 @@ import kotlin.properties.ReadOnlyProperty
 import kotlin.reflect.KClass
 import kotlin.reflect.KProperty
 
-open class App(primaryView: KClass<out UIComponent>? = null, vararg stylesheet: KClass<out Stylesheet>) : Application(), Configurable {
+open class App(open val primaryView: KClass<out UIComponent> = NoPrimaryViewSpecified::class, vararg stylesheet: KClass<out Stylesheet>) : Application(), Configurable {
     var scope: Scope = DefaultScope
     val workspace: Workspace get() = scope.workspace
 
@@ -52,17 +51,13 @@ open class App(primaryView: KClass<out UIComponent>? = null, vararg stylesheet: 
         FX.eventbus.fire(event)
     }
 
-    constructor(primaryView: KClass<out UIComponent>? = null, stylesheet: KClass<out Stylesheet>, scope: Scope = DefaultScope) : this(primaryView, *arrayOf(stylesheet)) {
+    constructor(primaryView: KClass<out UIComponent> = NoPrimaryViewSpecified::class, stylesheet: KClass<out Stylesheet>, scope: Scope = DefaultScope) : this(primaryView, *arrayOf(stylesheet)) {
         this.scope = scope
     }
 
-    constructor(icon: Image, primaryView: KClass<out UIComponent>? = null, vararg stylesheet: KClass<out Stylesheet>) : this(primaryView, *stylesheet) {
+    constructor(icon: Image, primaryView: KClass<out UIComponent> = NoPrimaryViewSpecified::class, vararg stylesheet: KClass<out Stylesheet>) : this(primaryView, *stylesheet) {
         addStageIcon(icon, scope)
     }
-
-    constructor() : this(null)
-
-    open val primaryView: KClass<out UIComponent> = primaryView ?: DeterminedByParameter::class
 
     fun <T : Any> k(javaClass: Class<T>): KClass<T> = javaClass.kotlin
 
@@ -119,14 +114,14 @@ open class App(primaryView: KClass<out UIComponent>? = null, vararg stylesheet: 
         }
     }
 
-    open fun shouldShowPrimaryStage() = true
+    open fun shouldShowPrimaryStage() = primaryView != NoPrimaryViewSpecified::class
 
     open fun createPrimaryScene(view: UIComponent) = Scene(view.getRootWrapper())
 
     @Suppress("UNCHECKED_CAST")
     private fun determinePrimaryView(): KClass<out UIComponent> {
-        if (primaryView == DeterminedByParameter::class) {
-            val viewClassName = requireNotNull(parameters.named?.get("view-class")) { "No provided --view-class parameter and primaryView was not overridden. Choose one strategy to specify the primary View" }
+        if (primaryView == NoPrimaryViewSpecified::class) {
+            val viewClassName = parameters.named?.get("view-class") ?: return NoPrimaryViewSpecified::class
             val viewClass = Class.forName(viewClassName)
 
             require(UIComponent::class.java.isAssignableFrom(viewClass)) { "Class specified by --class-name is not a subclass of tornadofx.View" }
@@ -139,10 +134,6 @@ open class App(primaryView: KClass<out UIComponent>? = null, vararg stylesheet: 
     @Suppress("UNCHECKED_CAST")
     inline fun <reified T> inject(scope: Scope = DefaultScope): ReadOnlyProperty<App, T> where T : Component, T : ScopedInstance = object : ReadOnlyProperty<App, T> {
         override fun getValue(thisRef: App, property: KProperty<*>) = find<T>(scope)
-    }
-
-    class DeterminedByParameter : View() {
-        override val root = Pane()
     }
 
     fun trayicon(image: BufferedImage, tooltip: String?, implicitExit: Boolean = false, autoSize: Boolean = false, op: TrayIcon.() -> Unit) {
@@ -201,6 +192,22 @@ open class App(primaryView: KClass<out UIComponent>? = null, vararg stylesheet: 
         }
     }
 
+}
+
+/**
+ * This is the default primary view parameter. It is used to signal that there
+ * is no primary view configured for the application and so the App start process
+ * should not show the primary stage upon startup, unless the --view-class parameter
+ * was passed on the command line.
+ *
+ * If no primary view is shown, the developer must use either the start() hook or some
+ * other means of determinining that the application has started. This would be good
+ * for applications where the default view depends upon some state, or where the app
+ * simply starts with a tray icon.
+ *
+ */
+class NoPrimaryViewSpecified : View() {
+    override val root = stackpane()
 }
 
 inline fun <reified T : Application> launch(vararg args: String) = Application.launch(T::class.java, *args)